requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorama==0.4.6
dash==2.18.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.11.2
Faker==33.0.0
Flask==3.0.3
fsspec==2024.10.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
locket==1.0.0
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
partd==1.4.2
plotly==5.24.1
py4j==0.10.9.7
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
tenacity==9.0.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.21.0
